feat: make prebuilt libs (#112)
* feat(android): make prebuilt libs and provide rnllamaBuildFromSource project prop

* feat(cpp): split header file (#113)

* feat(ios): make xcframework (wip)

* fix(cpp): avoid template usage

* fix(ios): build for platform

* fix(ios): correct framework build

* feat(ios): include metallib

* feat(ios): refactor build-ios.sh

* chore(ios, example): set RNLLAMA_BUILD_FROM_SOURCE env

* fix(android): build-arm64

* ci: add more build jobs

* fix(ci): build-ios-frameworks

* fix(cpp): api changes

* chore: format

* docs(readme): update

* docs(readme): update

* fix(ios): add missing props
