+++ example/android/gradle.properties
@@ -42,3 +42,7 @@ newArchEnabled=true
 # Use this property to enable or disable the Hermes JS engine.
 # If set to false, you will be using JSC instead.
 hermesEnabled=true
+
+# Use this property to enable or disable the RNLLAMA build from source.
+# If set to true, the RNLLAMA library will be built from source.
+rnllamaBuildFromSource=true
