+++ example/ios/.xcode.env.local
@@ -1,1 +1,1 @@
-export NODE_BINARY=/var/folders/4z/1d45cfts3936kdm7v9jl349r0000gn/T/yarn--1737601800519-0.4653948355776887/node
+export NODE_BINARY=/var/folders/4z/1d45cfts3936kdm7v9jl349r0000gn/T/yarn--1737683137807-0.2592527908357527/node
