+++ example/ios/Podfile
@@ -18,6 +18,8 @@ end
 
 ENV['RCT_NEW_ARCH_ENABLED'] = '1'
 
+ENV['RNLLAMA_BUILD_FROM_SOURCE'] = ENV['RNLLAMA_BUILD_FROM_SOURCE'] || '1'
+
 target 'RNLlamaExample' do
   config = use_native_modules!
