+++ ios/RNLlama.h
@@ -1,5 +1,9 @@
 #ifdef __cplusplus
-#import "rn-llama.hpp"
+#if RNLLAMA_BUILD_FROM_SOURCE
+#import "rn-llama.h"
+#else
+#import <rnllama/rn-llama.h>
+#endif
 #endif
 
 #import <React/RCTEventEmitter.h>
