+++ scripts/bootstrap.sh
@@ -227,12 +227,12 @@ patch -p0 -d ./cpp < ./scripts/ggml-quants.c.patch
 patch -p0 -d ./cpp < ./scripts/llama-mmap.cpp.patch
 
 if [ "$OS" = "Darwin" ]; then
-  # Build metallib (~1.4MB)
+  # Build metallib (~2.6MB)
   cd llama.cpp/ggml/src/ggml-metal
   xcrun --sdk iphoneos metal -c ggml-metal.metal -o ggml-metal.air
   xcrun --sdk iphoneos metallib ggml-metal.air   -o ggml-llama.metallib
   rm ggml-metal.air
-  cp ./ggml-llama.metallib ../../../../cpp/ggml-llama.metallib
+  mv ./ggml-llama.metallib ../../../../cpp/ggml-llama.metallib
 
   cd -
